Use same TestError in starknet-server/tests
