+++ crates/starknet-server/tests/common/mod.rs
@@ -1,4 +1,5 @@
 pub mod background_anvil;
 pub mod background_devnet;
 pub mod constants;
+pub mod errors;
 pub mod utils;
